refactor: renamed cmd dir
