+++ internal/cmd/extensions/cmd.go
@@ -13,26 +13,28 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package extensibility
+package extensions
 
 import (
-	"strings"
-
-	"github.com/mia-platform/miactl/internal/resources/extensibility"
-	"github.com/mia-platform/miactl/internal/table"
+	"github.com/mia-platform/miactl/internal/clioptions"
+	"github.com/spf13/cobra"
 )
 
-func printExtensionsList(extensions []*extensibility.Extension) string {
-	strBuilder := &strings.Builder{}
-	table := table.New(strBuilder)
-	table.Header("ID", "Name", "Description")
-	for _, extension := range extensions {
-		table.Row(
-			extension.ExtensionID,
-			extension.Name,
-			extension.Description,
-		)
+func CMD(o *clioptions.CLIOptions) *cobra.Command {
+	cmd := &cobra.Command{
+		Use:   "extensions",
+		Short: "Manage registered extensions",
+		Long:  "Manage registered extensions for the company.",
 	}
-	table.Print()
-	return strBuilder.String()
+
+	flags := cmd.PersistentFlags()
+	o.AddConnectionFlags(flags)
+	o.AddContextFlags(flags)
+	o.AddCompanyFlags(flags)
+	o.AddProjectFlags(flags)
+
+	cmd.AddCommand(
+		ListCmd(o),
+	)
+	return cmd
 }
