+++ internal/cmd/extensions/cmd_test.go
@@ -13,38 +13,18 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package extensibility
+package extensions
 
 import (
 	"testing"
 
-	"github.com/mia-platform/miactl/internal/resources/extensibility"
+	"github.com/mia-platform/miactl/internal/clioptions"
+
 	"github.com/stretchr/testify/require"
 )
 
-func TestPrintExtensionsList(t *testing.T) {
-	data := []*extensibility.Extension{
-		{
-			ExtensionID: "ext-1",
-			Name:        "Extension 1",
-			Description: "Description 1",
-		},
-		{
-			ExtensionID: "ext-2",
-			Name:        "Extension 2",
-			Description: "Description 2",
-		},
-	}
-
-	table := printExtensionsList(data)
-
-	expectedTokens := []string{
-		"ID", "NAME", "DESCRIPTION",
-		"ext-1", "Extension 1", "Description 1",
-		"ext-2", "Extension 2", "Description 2",
-	}
-
-	for _, expected := range expectedTokens {
-		require.Contains(t, table, expected)
-	}
+func TestRootCommandBuilder(t *testing.T) {
+	opts := clioptions.NewCLIOptions()
+	cmd := CMD(opts)
+	require.NotNil(t, cmd)
 }
